A folder for previous versions of the guide
The folder is for previous versions of the guide. It's mostly for keeping branches consistent, but if you want to download the older versions, you can do that as well. WARNING: Older versions of the guide might only be incompatible if Among Us has any major updates. This is not guaranteed, as I can't see the future and I don't have a time machine, but using the newest and most recent version is recommended, as it will be consistent with any new updates. However, I'm not the boss, so do whatever you want.